% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azureB2C-package.R
\docType{package}
\name{azureB2C-package}
\alias{azureB2C}
\alias{azureB2C-package}
\title{azureB2C: New slick project!}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Steph Locke \email{steph@itsalocke.com} (0000-0002-2387-3723)

}
\keyword{internal}
