% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2c_app.R
\name{b2c_app}
\alias{b2c_app}
\title{Create an Azure B2C Oauth application}
\usage{
b2c_app(app_name, client_id, client_secret = NULL,
  redirect_uri = "https://localhost:1410")
}
\arguments{
\item{app_name}{The (case-sensitve) name of the app as registered on Azure}

\item{client_id}{The client ID of the app}

\item{client_secret}{The (optional) client Secret of the app}

\item{redirect_uri}{The URL that the users should be returned to}
}
\value{
A list of class \code{oauth_app}
}
\description{
Create an Azure B2C Oauth application
}
\examples{
b2c_app("example","guid")
}
