% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2c_create_config.R
\name{b2c_create_config}
\alias{b2c_create_config}
\title{Create a starter config file}
\usage{
b2c_create_config(path = ".", ...)
}
\arguments{
\item{path}{Destination directory for config.yml}

\item{...}{Additional options for \code{yaml::write_yaml()}}
}
\description{
Create a starter config file
}
\examples{
\donttest{b2c_create_config()}
}
