% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2c_endpoint.R
\name{b2c_endpoint}
\alias{b2c_endpoint}
\title{The Azure AD B2C endpoint needed for authentication}
\usage{
b2c_endpoint(tenant, flow)
}
\arguments{
\item{tenant}{The Azure AD B2C tenant}

\item{flow}{The specific authentication workflow the user needs to route through}

\item{...}{Additional \code{httr::oauth_endpoint()} parameters}
}
\value{
A list of relevent endpoints
}
\description{
The Azure AD B2C endpoint needed for authentication
}
\examples{
b2c_endpoint("itsalocke", "login")
}
