% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2c_get_config.R
\name{b2c_get_config}
\alias{b2c_get_config}
\title{Get B2C config based on environment using the {config} package}
\usage{
b2c_get_config(path = ".", config = Sys.getenv("R_CONFIG_ACTIVE",
  "default"), ...)
}
\arguments{
\item{path}{Directory where config.yml is kept}

\item{config}{The name of the configuration needed}

\item{...}{Additional parameters for \code{config::get()}}
}
\value{
Config list object
}
\description{
Get B2C config based on environment using the {config} package
}
\examples{
\donttest{b2c_get_config()}
}
