% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2c_get_token.R
\name{b2c_get_token}
\alias{b2c_get_token}
\title{Acquire a token}
\usage{
b2c_get_token(config = list(app_name = "lddemo", client_id =
  "0e490a89-b382-4d3b-9f6b-60630dfd624b", client_secret = NULL,
  resource_uri = "https://jwt.ms", redirect_uri = "https://jwt.ms", scopes
  = "openid", tenant = "nightingalehq", flow = "B2C_1_SignUpAndSignIn"),
  ...)
}
\arguments{
\item{config}{B2C config, usually retrieved with \code{b2c_get_config()}}

\item{...}{Other arguments passed to \code{httr::oauth2.0_token()}}
}
\value{
a httr oauth2.0 token
}
\description{
Acquire a token
}
\examples{
\donttest{b2c_get_token()}
}
